Revert "Format changes and comments added"
