--- .gitignore
@@ -1,18 +1,0 @@
-# no .a files
-*.a
-*.iml
-
-# but do track lib.a, even though you're ignoring .a files above
-!lib.a
-
-# only ignore the root TODO file, not subdir/TODO
-/TODO
-
-# ignore all files in the build/ directory
-build/
-
-# ignore doc/notes.txt, but not doc/server/arch.txt
-doc/*.txt
-
-# ignore all .txt files in the doc/ directory
-doc/**/*.txt
