--- notes.txt
@@ -1,1 +1,0 @@
-1.Building web server for web services.
