+++ pom.xml
@@ -50,7 +50,6 @@
 			<groupId>org.springframework</groupId>
 			<artifactId>spring-test</artifactId>
 			<version>4.1.5.RELEASE</version>
-            <scope>test</scope>
 		</dependency>
