+++ retail-crm-services/src/main/java/org/crm/dao/model/OrderItemDao.java
@@ -6,10 +6,6 @@ import org.crm.entities.OrderItem;
 
 public interface OrderItemDao {
 
-    /**
-     *
-     * @param orderId the unique id for each order.
-     * @return a list of order item.
-     */
+
 	public List<OrderItem> findOrderItemsByOrderId(Integer orderId);
 }
