+++ retail-crm-services/src/main/java/org/crm/dao/model/ProductDao.java
@@ -2,15 +2,12 @@ package org.crm.dao.model;
 
 import org.crm.entities.Product;
 
-/**
- * The Product DAO interface.
- */
 public interface ProductDao {
 	
 	/**
 	 * 
-	 * @param productName the name of the production.
-	 * @return Product object.
+	 * @param productName
+	 * @return
 	 */
 	public Product getProductByName(String productName);
 }
